title section click and hover behavior, move useObserveHeights to another folder
