+++ packages/main/src/components/DynamicPageHeader/index.tsx
@@ -1,14 +1,15 @@
-import { createUseStyles } from 'react-jss';
+import { isIE } from '@ui5/webcomponents-react-base/dist/Device';
 import { StyleClassHelper } from '@ui5/webcomponents-react-base/dist/StyleClassHelper';
 import { usePassThroughHtmlProps } from '@ui5/webcomponents-react-base/dist/usePassThroughHtmlProps';
-import { FlexBox } from '@ui5/webcomponents-react/dist/FlexBox';
-import { FlexBoxAlignItems } from '@ui5/webcomponents-react/dist/FlexBoxAlignItems';
-import React, { FC, forwardRef, ReactNode, ReactNodeArray, Ref, useMemo } from 'react';
 import { CommonProps } from '@ui5/webcomponents-react/interfaces/CommonProps';
+import React, { FC, forwardRef, ReactNode, ReactNodeArray, Ref, useMemo } from 'react';
+import { createUseStyles } from 'react-jss';
 import { DynamicPageHeaderStyles } from './DynamicPageHeader.jss';
-import { isIE } from '@ui5/webcomponents-react-base/dist/Device';
 
 export interface DynamicPageHeaderProps extends CommonProps {
+  /**
+   * Content of the `DynamicPageHeader`.
+   */
   children?: ReactNode | ReactNodeArray;
 }
